requirements.txt
Flask==3.0.2
requests==2.31.0
supabase==2.3.4
APScheduler==3.10.4
python-dotenv==1.0.1
gunicorn==21.2.0
